Add needed classes to better implement MVC. Move action listeners to their own class.
