+++ src/main/java/UML/views/CommandlineView.java
@@ -1,0 +1,4 @@
+package UML.views;
+public class CommandlineView {
+    
+}
