+++ src/main/java/UML/views/GrahpicalView.java
@@ -1,0 +1,5 @@
+package UML.views;
+
+public class GrahpicalView {
+    
+}
